feat: optimize project and fix nested metric extraction
Major improvements:
- Enable incremental compilation via .cargo/config.toml
- Fix nested metric extraction in UnifiedExtractor
- Fix nested metric display in DeviceDetail component
- Reduce unwrap() usage with proper error handling
- Clean up redundant docs and expired test files

Backend fixes:
- unified_extractor.rs: Allow null values in extract_by_path()
- unified_extractor.rs: Add recursive expansion for nested objects
- auth_users.rs: Add create_hmac() helper for safe JWT signing
- timeseries.rs: Fix clippy never_loop warning

Frontend fixes:
- DeviceDetail.tsx: Add expandNestedValues() for nested metrics
- Display values.*, data.*, etc. as expanded metrics

Cleanup:
- Delete 20+ analysis reports from docs/
- Delete docs/archive/ directory
- Delete 28 expired evaluation test files
- Remove HASS discovery modules (deprecated)

+++ crates/commands/src/queue.rs
@@ -36,7 +36,7 @@ struct QueueItem {
 impl QueueItem {
     fn new(command: CommandRequest, sequence: u64) -> Self {
         // Priority is inverted for min-heap
-        let priority = (6 - command.priority.value()) as u8;
+        let priority = (6 - command.priority.value());
         Self {
             command,
             priority,
